requirements.txt
decorator==4.3.0
funcsigs==1.0.2
future==0.16.0
imageio==2.3.0
imutils==0.4.6
matplotlib==2.2.2
more-itertools==4.2.0
moviepy==0.2.3.5
numpy==1.14.5
numpy-stl==2.8.0
opencv-python
Pillow==5.2.0
pyquaternion==0.9.2
ray==0.6.0
scikit-image==0.14.0
scipy==1.1.0
six==1.11.0
tensorboard==1.6.0
tensorflow
mujoco_py==1.50.1.56
requests==2.20.0
h5py==2.9.0
torch==1.5.0
kornia==0.3.1
torchvision==0.6.0
tensorboardX==1.2.0
faiss
